Big updates to NeuralData class
The class definition and constructor were updated (in NeuralData.m).
The constructor now takes a string that describes the location of the
data (this should probably be updated to make more sense and then be
documented).

The functionality of the `DetectRipples` function was moved into the
method `detectRipples`.  Similarly, functions for computing
spectrograms, sharp-wave signals, and ripple-wave signals were moved
into methods (`getRippleSpectrogram`, `getRippleWave`, `getSharpWave`);
these methods automatically save their data based on the location of the
data.

The following helper methods were added:
* highLfp
* lowLfp
* mainLfp
* numDataChannels
* rawSampleRate
* setCurrentChannels
* loadChannels