`NeuralData`: use `matfile` for BehavElectrDataLFP
COULD POTENTIALLY CAUSE ISSUES

This data takes a long time to load, and a lot of it isn't even used.
This modification allows us to load the essential parts on demand.
Unfortunately, not all data structures can be accessed this way, which
could potentially cause unforseen issues. Because this commit only
modifies how these 5 property values are accessed, it should be safe to
revert it without causing any isses.